✨ v2.4.0 Add `ValkeyConnector` + bugfixes and improvements. (#167)
# v2.4.0

- **Add `valkey` instance connectors.**  
  Introducing a new first-class instance connector: the `ValkeyConnector`. [Valkey](https://valkey.io/), a fork of Redis, is a high-performance in-memory database often used for caching. The `valkey` service has been added to the Meerschaum stack and is accessible via the built-in connector `valkey:main`.

- **Cache Web Console sessions in Valkey when running with `--production`.**  
  Starting the web API with `--production` will now store sessions in `valkey:main`. This results in a smoother experience in the event of a web server restart. By default, sessions expire after 30 days.
  > You may disable this behavior by setting `system:experimental:valkey_session_cache` to `false`.

- **Allow for a default executor.**  
  Setting the key `meerschaum:executor` will set the default executor (overriding the check for `systemd`). This is useful for defaulting to remote actions in a multi-node deployment.

- **Allow querying for `None` in `query_df()`.**  
  You may now query for null rows:

  ```python
  import pandas as pd
  from meerschaum.utils.dataframe import query_df

  df = pd.DataFrame({'a': [1, 2, pd.NA]}).astype('Int64')

  result = query_df(df, {'a': None})
  print(result)
  #       a
  # 2  <NA>
  
  result = query_df(df, {'a': [None, 1]})
  print(result)
  #       a
  # 0     1
  # 2  <NA> 
  ```

- **Improve `query_df()` performance.**  
  Dataframe vlues are no longer serialized by default in `query_df()`, meaning that parameters must match the data type. Pass `coerce_types=True` to restore legacy behavior.

- **Add `Pipe.copy_to()`.**  
  Copy pipes between instances with `Pipe.copy_to()`:

  ```python
  import meerschaum as mrsm

  pipe = mrsm.Pipe('plugin:noaa', 'weather')
  pipe.copy_to('valkey:main')
  ```

- **Add `include_unchanged_columns` to `Pipe.filter_existing()`.**  
  Pass `include_unchanged_columns=True` to return entire documents in the update dataframe. This is useful for situations where you are unable to update individual fields:

  ```python
  import meerschaum as mrsm
  import pandas as pd

  pipe = mrsm.Pipe('a', 'b', columns=['id'])
  pipe.sync([
      {'animal': 'cat', 'name': 'Meowth', 'id': 1},
      {'animal': 'dog', 'name': 'Fluffy', 'id': 2},
  ])

  df = pd.DataFrame([{'id': 1, 'breed': 'tabby'}])

  unseen, update, delta = pipe.filter_existing(df)
  print(update)
  #    id  breed
  # 0   1  tabby
  
  unseen, update, delta = pipe.filter_existing(
      df,
      include_unchanged_columns=True,
  )
  print(update)
  #   animal    name  id  breed
  # 0    cat  Meowth   1  tabby
  ```

- **Add a share button to the Pipe card.**  
  On the web dashboard, you may now more easily share pipes by clicking the "share" icon and copying the URL. This opens the pipe card in a new, dedicated tab.

- **Add `OPTIONAL_ATTRIBUTES` to connectors.**  
  Connectors may now set `OPTIONAL_ATTRIBUTES`, which will add skippable prompts in `bootstrap connector`.

- **Remove progress bar for syncing via remote actions.**  
  Executing `sync pipes` remotely will no longer print the timer progress bar.

- **Fix bug with `stack` in the shell.**  
  Note that `stack` actions may not be chained.

- **Fix scheduler dependency.**  
  To fix the installation of `APScheduler`, `attrs` is now held back to 24.1.0.

+++ tests/test_pipes_dtypes.py
@@ -10,7 +10,6 @@ from tests.connectors import conns, get_flavors
 from tests.test_users import test_register_user
 import meerschaum as mrsm
 from meerschaum import Pipe
-from meerschaum.actions import actions
 from meerschaum.utils.dtypes import are_dtypes_equal
 from meerschaum.utils.sql import sql_item_name
 
@@ -18,6 +17,7 @@ from meerschaum.utils.sql import sql_item_name
 def run_before_and_after(flavor: str):
     test_register_user(flavor)
     yield
+
 
 @pytest.mark.parametrize("flavor", get_flavors())
 def test_sync_change_columns_dtypes(flavor: str):
@@ -54,11 +54,11 @@ def test_dtype_enforcement(flavor: str):
     pipe.delete(debug=debug)
     pipe = Pipe(
         'dtype', 'enforcement',
-        columns = {
+        columns={
             'datetime': 'dt',
             'id': 'id',
         },
-        dtypes = {
+        dtypes={
             'int': 'int',
             'float': 'float',
             'bool': 'bool',
@@ -67,7 +67,7 @@ def test_dtype_enforcement(flavor: str):
             'numeric': 'numeric',
             'str': 'str',
         },
-        instance = conn,
+        instance=conn,
     )
     pipe.sync([{'dt': '2022-01-01', 'id': 1, 'int': '1'}], debug=debug)
     pipe.sync([{'dt': '2022-01-01', 'id': 1, 'float': '1.0'}], debug=debug)
@@ -448,24 +448,26 @@ def test_sync_bools_inplace(flavor: str):
     Test that pipes are able to sync bool in-place.
     """
     conn = conns[flavor]
+    if conn.type not in ('api', 'sql'):
+        return
     pipe = mrsm.Pipe('test', 'bools', 'inplace', instance=conn)
     _ = pipe.delete()
     pipe = mrsm.Pipe(
         'test', 'bools', 'inplace',
-        instance = conn,
-        columns = {'datetime': 'dt', 'id': 'id'},
+        instance=conn,
+        columns={'datetime': 'dt', 'id': 'id'},
     )
     pipe_table = sql_item_name(pipe.target, conn.flavor) if conn.type == 'sql' else pipe.target
     inplace_pipe = mrsm.Pipe(conn, 'bools', 'inplace', instance=conn)
     _ = inplace_pipe.delete()
     inplace_pipe = mrsm.Pipe(
         conn, 'bools', 'inplace',
-        instance = conn,
-        columns = pipe.columns,
-        dtypes = {
+        instance=conn,
+        columns=pipe.columns,
+        dtypes={
             'is_bool': 'bool',
         },
-        parameters = {
+        parameters={
             'fetch': {
                 'definition': f"SELECT * FROM {pipe_table}",
                 'pipe': pipe.keys(),
@@ -500,4 +502,3 @@ def test_sync_bools_inplace(flavor: str):
     assert success, msg
     df = inplace_pipe.get_data(params={'id': 3})
     assert 'na' in str(df['is_bool'][0]).lower()
-
